+++ storage/types/entry.js
@@ -36,11 +36,11 @@ module.exports = exports = class Entry {
     if (options.tags) this.tags = options.tags;
   }
 
-  get smt_id() {
+  get smt_urn() {
     if (this.domain)
-      return this.domain + "/" + this.name;
+      return this.domain + ":" + this.name;
     else
-      return this.name;
+      return ":" + this.name;
   }
 
 };