+++ test/codex/codex_in-memory.js
@@ -47,12 +47,12 @@ async function test(schema) {
     logger.verbose(JSON.stringify(results, null, "  "));
 
     // recall encoding
-    let smt_id = entry.smt_id;
-    results = await Storage.codex.recall(smt_id);
+    let smt_urn = entry.smt_urn;
+    results = await Storage.codex.recall(smt_urn);
     logger.verbose("recall: " + results.resultText);
 
     if (results.resultCode === 0) {
-      encoding = results.data[ smt_id ];
+      encoding = results.data[ smt_urn ];
 
       let outputfile = "./data/output/codex/" + schema + ".encoding.json";
       logger.verbose("output file: " + outputfile);